LayoutTests/media/media-source/media-source-changetype-vp8-vp9-expected.txt


EVENT(update)
EVENT(update)
EVENT(update)
EVENT(update)
EXPECTED (video.currentTime >= '0.6') OK
END OF TEST
